Disconnect SFTP channel when inputstream is closed (#1320)
